Enable encounters in quest mode
Due to PD being able to hide Pokemon on the game map, it is no longer a
problem that Pokemon might be clicked instead of the Pokestop during
quest scan. Thus, checking for these clicks based on received encounter
data is no longer necessary and encounters can be received without
disturbing the quest scan flow.